Refactor shell quoting
Move the shlex-like quoting utils in the uv-shell crate
