+++ crates/uv/src/commands/venv.rs
@@ -25,7 +25,7 @@ use uv_python::{
 };
 use uv_resolver::{ExcludeNewer, FlatIndex};
 use uv_settings::PythonInstallMirrors;
-use uv_shell::Shell;
+use uv_shell::{shlex_posix, shlex_windows, Shell};
 use uv_types::{BuildContext, BuildIsolation, HashStrategy};
 use uv_warnings::{warn_user, warn_user_once};
 use uv_workspace::{DiscoveryOptions, VirtualProject, WorkspaceError};
@@ -420,38 +420,4 @@ async fn venv_impl(
     }
 
     Ok(ExitStatus::Success)
-}
-
-/// Quote a path, if necessary, for safe use in a POSIX-compatible shell command.
-fn shlex_posix(executable: impl AsRef<Path>) -> String {
-    // Convert to a display path.
-    let executable = executable.as_ref().portable_display().to_string();
-
-    // Like Python's `shlex.quote`:
-    // > Use single quotes, and put single quotes into double quotes
-    // > The string $'b is then quoted as '$'"'"'b'
-    if executable.contains(' ') {
-        format!("'{}'", executable.replace('\'', r#"'"'"'"#))
-    } else {
-        executable
-    }
-}
-
-/// Quote a path, if necessary, for safe use in `PowerShell` and `cmd`.
-fn shlex_windows(executable: impl AsRef<Path>, shell: Shell) -> String {
-    // Convert to a display path.
-    let executable = executable.as_ref().user_display().to_string();
-
-    // Wrap the executable in quotes (and a `&` invocation on PowerShell), if it contains spaces.
-    if executable.contains(' ') {
-        if shell == Shell::Powershell {
-            // For PowerShell, wrap in a `&` invocation.
-            format!("& \"{executable}\"")
-        } else {
-            // Otherwise, assume `cmd`, which doesn't need the `&`.
-            format!("\"{executable}\"")
-        }
-    } else {
-        executable
-    }
 }
